@ralfbrown 2020-12-27
Weird.  SIMD_FOR is just the equivalent of writing
```
    #ifdef _OPENMP
    #pragma omp simd
   #endif
   for
```
So what your change does is merely remove the hint to the compiler that the loop is vectorizable....   I'd love to see the generated assembly for the two cases to figure out what's going on there.