@n0s0r0g 2020-12-27
`for` loop:
```
Address Source Line Assembly
0x144184	481	vmulss xmm4, xmm0, dword ptr [r8]
0x144189	481	vmulss xmm3, xmm0, xmm3
0x14418d	481	vmulss xmm1, xmm0, xmm1
0x144191	481	vaddss xmm0, xmm0, dword ptr [rcx+0xc]
0x144196	481	add rcx, 0x10
0x14419a	481	vaddss xmm4, xmm4, dword ptr [rcx-0x10]
0x14419f	481	vaddss xmm3, xmm3, dword ptr [rcx-0xc]
0x1441a4	481	vaddss xmm1, xmm1, dword ptr [rcx-0x8]
0x1441a9	481	vmovss dword ptr [rcx-0x10], xmm4
0x1441ae	481	vmovss dword ptr [rcx-0xc], xmm3
0x1441b3	481	vmovss dword ptr [rcx-0x8], xmm1
0x1441b8	481	vmovss dword ptr [rcx-0x4], xmm0
```

`SIMD_FOR` loop:
```
Address	Source Line	Assembly
0x14418b	478	mov rax, r8				
0x14418e	478	sub rax, r9				
0x144191	478	vmovss xmm1, dword ptr [rax+r14*1]				
0x144197	478	mov dword ptr [rbp-0x44], 0x3f800000				
0x14419e	478	vmovss dword ptr [rbp-0x50], xmm1				
0x1441a3	478	vmovss xmm1, dword ptr [r8]				
0x1441a8	481	vshufps xmm0, xmm0, xmm0, 0x0				
0x1441ad	478	vmovss dword ptr [rbp-0x4c], xmm1				
0x1441b2	478	vmovss xmm1, dword ptr [rax+rcx*1]				
0x1441b7	478	add rsi, 0x4				
0x1441bb	478	vmovss dword ptr [rbp-0x48], xmm1				
0x1441c0	481	vmulps xmm0, xmm0, xmmword ptr [rbp-0x50]				
0x1441c5	481	add r8, 0x10				
0x1441c9	481	add rdi, 0x10				
0x1441cd	481	vaddps xmm0, xmm0, xmmword ptr [rdi-0x10]				
0x1441d2	481	vmovups xmmword ptr [rdi-0x10], xmm0				
```