@ralfbrown 2020-12-27
So the compiler *is* generating vectorized code only in the case with the hint, but clearly the extra shuffling to get things where they need to be to use vector instructions outweighs the benefit.  Looks like the optimizer's cost model is incorrect for your CPU.