@n0s0r0g 2020-12-28
I guess the hint misleads the compiler that is it beneficial to vectorize. I can't blame the compiler for 2.5x faster loop code if we compare CPU time. Do you expect performance degradation on other CPU's?

By the way, `0020-denoise-nlmeans` test has the highest performance degratation with SSE2-optimized codepath disabled:
native: `2.71` sec (2.708; 2.729; 2.703)
native-no-sse2: `7.9` sec (7.927; 7.889; 7.897)