@aurelienpierre 2020-12-30
@ralfbrown The `SIMD_FOR` macro is not a good idea, in my opinion, since you need to additionnaly declare `aligned(buffer:16)` or `aligned(buffer:64)` for the compiler to actually make a clever use of the vectorization. I have also, in the past, witnessed degraded performance with `simd for` when the alignment is not declared (and it's referenced in many OpenMP tutos all around).