@n0s0r0g 2021-01-01
It looks like vectorized versions are slower due to "Loads Blocked by Store Forwarding" issue.