@ralfbrown 2021-01-01
OK, I see what's happening in those two cases -- the writes to the pixel[] array aren't being optimized away into a load of an xmm register, so we have four 32-bit memory writes in flight at the time of the 128-bit memory read that the loop got vectorized into.  What is the performance like if we simply move the setting of pixel[] _before_ the computation of `wt`?
```
              const float *const inpx = in+4*col;
              const float DT_ALIGNED_ARRAY pixel[4] = { inpx[offset],  inpx[offset+1], inpx[offset+2], 1.0f };
              distortion += (col_sums[col+radius] - col_sums[col-radius-1]);
              const float wt = gh(distortion * sharpness);
              SIMD_FOR (size_t c = 0; c < 4; c++)
              {
                out[4*col+c] += pixel[c] * wt;
              }
```
I just checked, and GCC8 leaves the stores to `pixel` before the `wt` computation.