@n0s0r0g 2021-01-01
Test: `0020-denoise-nlmeans`

`SIMD_FOR_new`: 7.476s (7.527; 7.443; 7.458)
`for_new`: 5.343s (5.331; 5.366; 5.332)

Previous measurements:
`SIMD_FOR`: 7.9s (7.927; 7.889; 7.897)
`for`: 5.3s (5.285; 5.337; 5.272)