@ralfbrown 2021-01-01
OK, it helps, but there obviously still isn't enough time between the last write to `pixel` until it gets read as a unit.
Will have to think about this.