@ralfbrown 2021-01-01
I've tried a number of variations, and found one that makes GCC8 use a vector register for pixels[4], but in all the variations I tried, my 3rd-gen Threadripper showed exactly the same time within measurement error (so it's not being affected by the Loads Blocked by Store Forwarding).  That's SIMD_FOR, plain `for`, rearranging the setup of `pixels` before the computation of `wt`, etc.

Here's what worked:
```
   const float DT_ALIGNED_PIXEL pixel[4] = { inpx[offset],  inpx[offset+1], inpx[offset+2], 1.0f };
#pragma omp simd aligned(pixel, out:16)
    for (size_t c = 0; c < 4; c++)
```
GCC's optimizer did miss out on an optimization, however: it loads pixel[] into a vector register, inserts the four floats, then stores it back.  But it then uses that vector register as one of the args to the multiplication, so it doesn't need a memory read.