@n0s0r0g 2021-01-01
Test: `0020-denoise-nlmeans`
`omp_simd_ver2`: 4.745 (4.747; 4.749; 4.740)

This version is faster.